Complete rewrite of the emulation layer.
The changes are too numerous to mention.
The terminal emulation layer has been completely rewritten, and
now is much more accurate in its emulation of various classic
terminals. Things are also orders of magnitude faster.

mtm also passes all of the VTTEST tests that I want it to
(the ones it doesn't pass now it probably never will, because
they're not meaningful for what mtm is trying to do).