Fix VT52/VT100 transition; add more handled keys.
Originally, mtm simply had two vtparsers: one for VT100 sequences,
and one for VT52 sequences. When transitioning between the two modes,
we'd simply swap parsers.

This was a problem, though, because we want to be able to write more
than 1 byte at a time to the parser for the sake of efficiency. What
happens if we write 100 bytes, but byte 38 effects a transition?
The remaining bytes will be written to the wrong parser.

Now we simply have one parser, and reset/rewire the parser in callbacks
wired up depending on mode.

This whole thing started, by the way, because we wanted to better
support more special keysyms, including VT52 arrow keys...so, as
an aside in this commit, some more keycodes were handled.