Set the window title in the only guaranteed-to-work way.
ncurses does not have a way to set the title on a terminal
supporting the XT capability (e.g. xterm). We can inject
stuff into ncurses's output stream with (e.g.) putp, but
it doesn't always work because of partially-constructed
command sequences and delays in flushing and so on. The only
completely "correct" way to do it is to drop out of ncurses
temporarily, set the title, and drop back in. It's ugly, but it
works.