Makefile-alt with CURSESCFLAGS and CURSESLIBS options
An alternate Makefile better suited to people using `pkg-config` for their ncurses library (the one on my Ubuntu 16.04 32bits is obsolete, it is still v5 while MTM expects v6). I could have added an invocation to `pkg-config`, but it makes it more general if it just use two `CURSESCFLAGS` and `CURSESLIBS`. Not `CURSESLIBS` is not always just a matter of paths, it may be a matter of linker options. Ex for me, `pkg-config` says:

    $ pkg-config --cflags ncursesw
    > -D_DEFAULT_SOURCE -D_XOPEN_SOURCE=600 -I/home/yannick/apps/ncurses/include/ncursesw -I/home/yannick/apps/ncurses/include
    $ pkg-config --libs ncursesw
    > -L/home/yannick/apps/ncurses/lib -Wl,-rpath,/home/yannick/apps/ncurses/lib -lncursesw

Note the `-Wl,-rpath …` part, which is why a library path is not always enough. For a custom ncurses library, an `-I…` option is also required for the C compilation.

This is why I’m proposing alternate Makefile addition, with `CURSESCFLAGS` and `CURSESLIBS` options.